Fixed some functions in clientmain + improved client's main ui + added reload button to client's main and admin's main + selection error handling + linked client signup w/ main

+++ clientmain.py
@@ -23,7 +23,7 @@ def getTrips(tree):
     countOfTrips=cursor.fetchone()
     cursor.execute("""SELECT TripID,TRIP.RouteID,PickupStation,PickupTime,AvailableSeats,TripStatus,TripFee 
                     FROM TRIP inner join ROUTE
-                    ON ROUTE.ROUTEID=TRIP.ROUTEID""")
+                    ON ROUTE.ROUTEID=TRIP.ROUTEID WHERE TRIPSTATUS='Not Completed'""")
     cell_colors=["cell1","cell2"]
     color_index=0
     tree.delete(*tree.get_children())
@@ -40,7 +40,7 @@ def getBookings(tree,clientID):
     cursor.execute("""SELECT COUNT(*) FROM BOOKS_A WHERE CLIENTID=?""",clientID)
     # cursor.execute("""SELECT MAX(TRIP.TRIPID) FROM TRIP""")
     countOfTrips=cursor.fetchone()
-    cursor.execute("""SELECT BOOKINGID,BOOKS_A.TRIPID,PICKUPSTATION,TRIP.ROUTEID,PickupTime,TripStatus,FIRSTNAME,LASTNAME,GENDER,LICENSEPLATE,TELEPHONENUM,TripFee,NUMBEROFSEATS,DRIVERRATING,TRIPRATING 
+    cursor.execute("""SELECT BOOKINGID,BOOKS_A.TRIPID,PICKUPSTATION,TRIP.ROUTEID,PickupTime,TripStatus,FIRSTNAME,LASTNAME,LICENSEPLATE,TELEPHONENUM,TripFee,NUMBEROFSEATS,DRIVERRATING,TRIPRATING 
                     FROM BOOKS_A inner join TRIP 
                     ON TRIP.TRIPID=BOOKS_A.TRIPID AND BOOKS_A.CLIENTID=?
                     inner join DRIVER
@@ -57,7 +57,7 @@ def getBookings(tree,clientID):
     for i in range(countOfTrips[0]):
             bookingData=cursor.fetchone()
             if(bookingData):
-                tree.insert('', 'end',text="1", values=(bookingData[0],bookingData[1],bookingData[2],bookingData[3],bookingData[4],bookingData[5],bookingData[6]+" "+bookingData[7],bookingData[8],bookingData[9],bookingData[10],bookingData[11],bookingData[12],bookingData[13],bookingData[14]), tags=(cell_colors[color_index]))
+                tree.insert('', 'end',text="1", values=(bookingData[0],bookingData[1],bookingData[2],bookingData[3],bookingData[4],bookingData[5],bookingData[6]+" "+bookingData[7],bookingData[8],bookingData[9],bookingData[10],bookingData[11],bookingData[12],bookingData[13]), tags=(cell_colors[color_index]))
             if color_index==0:
                     color_index+=1
             elif color_index==1:
@@ -66,6 +66,8 @@ def getBookings(tree,clientID):
 
 def bookTrip(tree,clientId,booking_id,numOfSeats):
     try:
+       if(len(tree.selection())==0):
+            raise ValueError("Please select a Trip")
        selected_item = tree.selection()[0]
        values = tree.item(selected_item,'values')
        isBooked = checkTripIsBooked(values[0],clientId)
@@ -86,29 +88,44 @@ def bookTrip(tree,clientId,booking_id,numOfSeats):
        else:
             raise ValueError("Cannot book a completed trip")
     except Exception as e:
-        print(e)
+        if(str(e)=="Cannot book a completed trip" or str(e)=="Trip already booked." or str(e)=="The available seats in this trip is not enough." or str(e)=="Please select a Trip"):
+            print(e)
+            messagebox.showinfo("Booking failed",e)
 
 def deleteBooking(tree,clientId):
     try:
+        if(len(tree.selection())==0):
+            raise ValueError("Please select a Trip Booking")
         selected_item = tree.selection()[0]
         values=tree.item(selected_item,'values')
         isCompleted = values[5]
         if(isCompleted != "Completed"):
             cursor.execute('DELETE FROM BOOKS_A WHERE BOOKINGID = ? AND TRIPID = ?',values[0],values[1])
             cursor.commit()
-            cursor.execute('UPDATE TRIP SET AVAILABLESEATS = AVAILABLESEATS + ? where TRIPID = ?',values[11],values[1])
+            cursor.execute('UPDATE TRIP SET AVAILABLESEATS = AVAILABLESEATS + ? where TRIPID = ?',values[10],values[1])
             cursor.commit()
             getBookings(tree,clientId)
         else:
-            raise ValueError("Cannot delete a completed trip")
+            raise ValueError("Cannot delete booking of a completed trip")
     except Exception as e:
-        print(e)
+        messagebox.showerror("Booking deletion failed",e)
 
 def giveRating(tree,clientId,driver_rate_entry,trip_rate_entry):
     try:
+        if(len(tree.selection())==0):
+            raise ValueError("Please select a completed trip of yours")
         selected_item = tree.selection()[0]
         values = tree.item(selected_item,'values')
         tripStatus = values[5]
+        # values[1] tripID
+        cursor.execute("SELECT COUNT(*) FROM (SELECT * FROM FEEDBACK WHERE CLIENTID=? AND TRIPID=?)as cr",clientId,values[1])
+        if(driver_rate_entry.get()==""or trip_rate_entry.get()==""):
+            raise ValueError("Please add a rating (1 to 5) for both")
+        if(cursor.fetchone()[0]>0):
+            raise ValueError("Already rated")
+        if(int(driver_rate_entry.get())>5 or int(trip_rate_entry.get())>5 or int(driver_rate_entry.get())<0 or int(trip_rate_entry.get())<0 ):
+            raise ValueError("Not a valid rating")
+            
         if(tripStatus == "Completed"):
             cursor.execute("""INSERT INTO FEEDBACK
                             VALUES(?,?,?,?)""",clientId,values[1],driver_rate_entry.get(),trip_rate_entry.get())
@@ -117,4 +134,8 @@ def giveRating(tree,clientId,driver_rate_entry,trip_rate_entry):
         else:   
             raise ValueError("Trip not completed!")
     except Exception as e:
-        print(e)
+        messagebox.showinfo("Failed to give a rating",e)
+        
+def reloadClientsMainPage(treeTrip,treeBooking,clientID):
+    getTrips(treeTrip)
+    getBookings(treeBooking,clientID)
